Move Login components up a directory
