+++ src/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { Route, withRouter, Switch } from 'react-router-dom';
 
 import RegistrationPage from './components/auth-components/registration-page';
-import LoginPage from './components/auth-components/login-page';
+import LoginPage from './components/LoginPage';
 import Lists from './components/Lists/Lists.js';
 import Items from './components/Items';
 import LandingPage from './components/LandingPage';
