+++ src/components/LoginForm/index.js
@@ -1,0 +1,3 @@
+import LoginForm from './LoginForm';
+
+export default LoginForm;
