+++ src/components/LoginPage/LoginPage.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
-import LoginForm from './login-form';
-import './auth.css';
+import LoginForm from '../LoginForm';
+import '../auth-components/auth.css';
 
 export function LoginPage(props) {
   // If we are logged in (which happens automatically when registration
